flower_delivery/requirements.txt
# Основные зависимости Django
Django==5.0
django-environ==0.11.2  # Для работы с переменными окружения

# Зависимости для работы с базой данных (если используется PostgreSQL)
psycopg2-binary==2.9.9

# Зависимости для Telegram бота
python-telegram-bot==20.6

# Зависимости для тестирования
coverage==7.3.2  # Для измерения покрытия кода тестами
factory-boy==3.3.0  # Для создания тестовых данных
Faker==19.6.1  # Для генерации случайных данных

# Дополнительные утилиты
python-dotenv==1.0.0  # Для загрузки переменных окружения из .env файла